Tokenizer/AttributesTest: move parse error test to own file
