+++ tests/Core/Tokenizers/PHP/AttributesParseError1Test.inc
@@ -1,0 +1,8 @@
+<?php
+
+// Intentional parse error.
+// This must be the only test in the file.
+
+/* testInvalidAttribute */
+#[ThisIsNotAnAttribute
+function invalid_attribute_test() {}
