+++ tests/Core/Tokenizers/PHP/AttributesParseError1Test.php
@@ -1,0 +1,39 @@
+<?php
+/**
+ * Tests the support of PHP 8 attributes
+ *
+ * @author    Alessandro Chitolina <alekitto@gmail.com>
+ * @copyright 2019 Squiz Pty Ltd (ABN 77 084 670 600)
+ * @license   https://github.com/PHPCSStandards/PHP_CodeSniffer/blob/HEAD/licence.txt BSD Licence
+ */
+
+namespace PHP_CodeSniffer\Tests\Core\Tokenizers\PHP;
+
+use PHP_CodeSniffer\Tests\Core\Tokenizers\AbstractTokenizerTestCase;
+
+final class AttributesParseError1Test extends AbstractTokenizerTestCase
+{
+
+
+    /**
+     * Test that invalid attribute (or comment starting with #[ and without ]) are parsed correctly.
+     *
+     * @covers PHP_CodeSniffer\Tokenizers\PHP::tokenize
+     * @covers PHP_CodeSniffer\Tokenizers\PHP::findCloser
+     * @covers PHP_CodeSniffer\Tokenizers\PHP::parsePhpAttribute
+     *
+     * @return void
+     */
+    public function testInvalidAttribute()
+    {
+        $tokens = $this->phpcsFile->getTokens();
+
+        $attribute = $this->getTargetToken('/* testInvalidAttribute */', T_ATTRIBUTE);
+
+        $this->assertArrayHasKey('attribute_closer', $tokens[$attribute]);
+        $this->assertNull($tokens[$attribute]['attribute_closer']);
+
+    }//end testInvalidAttribute()
+
+
+}//end class
